Basic layout proof of concept done
The layout is generated by a single js file using react and jsx. I would like to spit it up but could not figure out how.

It cannot be interacted with yet, but it has been built in a way that interactions and changing states shouldn't be impossible to implement